moved UserPage into pages
deleted unused components
